fix: 커스텀 hook 제거 및 관련 코드 복구
