--- frontend/src/hooks/member/index.ts
@@ -1,2 +1,0 @@
-export { useMemberProfiles } from './useMemberProfiles'
-export { useMemberCard } from './useMemberCard'
