+++ frontend/src/pages/member/default/MemberPage.tsx
@@ -1,11 +1,10 @@
-import { useEffect, useState } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 
 import { IntersectionObserverLoader } from '@/components/common'
 import { MemberCard } from '@/components/feature'
 import { PaginationNext, PaginationPrevious } from '@/components/ui'
-import { useMemberProfiles } from '@/hooks'
 import { ADMIN_PROFILE_MOCK } from '@/mock'
-import { getJoinSemestersApi } from '@/service/api'
+import { getJoinSemestersApi, useProfileSuspensePaging } from '@/service/api'
 
 export default function MemberPage() {
   const [joinSemesterData, setJoinSemesterData] = useState<string[]>([])
@@ -18,8 +17,15 @@ export default function MemberPage() {
     })
   }, [])
 
-  const { memberProfiles, hasNextPage, fetchNextPage, isFetchingNextPage } =
-    useMemberProfiles(joinSemesterData[semesterIndex])
+  const {
+    data: memberData,
+    hasNextPage,
+    isFetchingNextPage,
+    fetchNextPage,
+  } = useProfileSuspensePaging({
+    roles: ['ROLE_WEB_MASTER', 'ROLE_TEAM_LEADER', 'ROLE_MEMBER'],
+    joinSemester: joinSemesterData[semesterIndex],
+  })
 
   const handleLeftSemester = () => {
     if (semesterIndex > 0) {
@@ -41,8 +47,12 @@ export default function MemberPage() {
     return `${year}-${semester} 멤버`
   }
 
+  const memberProfiles = useMemo(() => {
+    return memberData?.pages.flatMap((page) => page.profiles) || []
+  }, [memberData])
+
   return (
-    <main className="flex h-full w-full flex-col items-center pb-20">
+    <main className="flex flex-col items-center pb-20 w-full h-full">
       <div className="w-full max-w-[920px] pb-4 text-xl font-semibold">
         해구르르
       </div>
@@ -60,7 +70,7 @@ export default function MemberPage() {
           )
         })}
       </div>
-      <div className="flex items-center gap-2 pt-10 text-xl font-semibold">
+      <div className="flex gap-2 items-center pt-10 text-xl font-semibold">
         <PaginationPrevious
           to="#"
           onClick={handleLeftSemester}
